refactor: consolidate all components into app/components
Move all shared components from root /components to /app/components
to centralize App Router related code. Update all import paths
accordingly.
