+++ app/components/page-detail.tsx
