+++ app/contact/components/contact-form.tsx
@@ -2,8 +2,8 @@
 
 import dynamic from "next/dynamic";
 import { useCallback, useEffect, useState } from "react";
-import { Page, type ListBlockChildrenResponseEx } from "../../components/rotion-wrappers";
-import Hed from "../../../components/hed";
+import { Page, type ListBlockChildrenResponseEx } from "../components/rotion-wrappers";
+import Hed from "../components/hed";
 
 // Dynamically import to load only on client-side
 const MutatingDots = dynamic(
