+++ app/not-found.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import Hed from "../components/hed";
+import Hed from "./components/hed";
 import { MakeOgImage } from "../src/lib/ogimage";
 
 const title = "404";
