+++ app/workout/components/workout-client.tsx
@@ -3,9 +3,9 @@
 import { Chart, type ChartData, registerables } from "chart.js";
 import { useState } from "react";
 import { Line } from "react-chartjs-2";
-import { Table, type FetchDatabaseRes } from "../../components/rotion-wrappers";
+import { Table, type FetchDatabaseRes } from "../components/rotion-wrappers";
 
-import Hed from "../../../components/hed";
+import Hed from "../components/hed";
 import Styles from "../../../styles/Workout.module.css";
 
 Chart.register(...registerables);
