+++ src/lib/ogimage.ts
@@ -4,7 +4,7 @@ import React from "react";
 import satori, { type SatoriOptions } from "satori";
 import sharp from "sharp";
 import twemoji from "twemoji";
-import OgImage from "../../components/ogimage";
+import OgImage from "../../app/components/ogimage";
 
 const notosansBold = readFileSync("./src/fonts/NotoSansJP-Black.woff");
 const notosansRegular = readFileSync("./src/fonts/NotoSansJP-Regular.woff");
